WIP #959 List project support for new data format
